refactor: address util 추가
api 호출로 받은 전체 주소를 동 포멧으로 필터링
